chore: use long maxTotalSize

+++ local-s3-core/src/main/java/com/robothy/s3/core/storage/Storage.java
@@ -34,7 +34,7 @@ public interface Storage {
    * @param maxTotalSize max total size.
    * @return a {@linkplain Storage} instance.
    */
-  static Storage createInMemory(int maxTotalSize) {
+  static Storage createInMemory(long maxTotalSize) {
     return new InMemoryStorage(maxTotalSize);
   }
